added folders, compose, draft and trash featchers

+++ src/pages/EmailIndex.jsx
@@ -1,8 +1,9 @@
 import { emailService } from "../services/email.service"
 import { MailList } from '../cmps/MailList'
 import { MailFilter } from '../cmps/MailFilter'
-import { MailDeteils } from "./MailDeteils";
+import { MailDeteils } from "./MailDeteils"
 import { Folders } from '../cmps/Folders'
+import { Compose } from '../cmps/compose'
 
 import imgUrl from '../assets/imgs/gmail-logo.jpg'
 
@@ -11,11 +12,14 @@ import { useEffect, useState} from "react"
 
 
 
+
 export function EmailIndex() {
     const [emails, setEmeils] = useState(null)
     const [ filterBy, setFilterBy ] = useState(emailService.defoultFilter())
     const [ clickChange, setClickChange ] = useState(false)
     const [ mode, setModeChange ] = useState("list")
+    const [ compose, setComposeChange ] = useState(false)
+    const [ comMail, setComposedMail ] = useState({})
 
     async function loadEmail() {
         const t = await emailService.query(filterBy)
@@ -33,6 +37,11 @@ export function EmailIndex() {
         console.log("change")
         loadEmail()
     }, [filterBy, clickChange])
+
+    useEffect(() => {
+        console.log("compose save" , comMail)
+        saveDraft()
+    }, [comMail])
 
     // function onFilterBy(filterBy){
     //     setFilterBy(filterBy)
@@ -59,17 +68,17 @@ export function EmailIndex() {
     function onHandleClick(id, action, ev){
        // ev.stopPropagation()
         setAction(id, action)
-        console.log("click",id, action)
+        // console.log("click",id, action)
     }
 
     async function setAction(id, action) { 
         if(action === "list") return setModeChange("list") 
 
         var t = await emailService.getById(id)
-        console.log(action)
         if(action === "read") t.isRead = true
         if(action === "star") t.isStarred = !t.isStarred
         if(action === "deteils") {
+            if(t.status === "draft") return onEdit(t)
             t.isRead = true
             setModeChange("deteils") 
         }
@@ -81,12 +90,56 @@ export function EmailIndex() {
          })
 
          if(action === "delete" || action === "delete-from-deteils"){
-         emailService.remove(id)
-         .then(() => {
-             setClickChange(prev => !prev)
-             if(action === "delete-from-deteils") setModeChange("list") 
-          })
+
+            if (t.status === "trash") { 
+
+                emailService.remove(id)
+                .then(() => {
+                    setClickChange(prev => !prev)
+                    if(action === "delete-from-deteils") setModeChange("list") 
+                })
+            }
+            else{
+                // console.log()
+                t.status = "trash"
+                await emailService.save(t)
+                if(action === "delete-from-deteils") setModeChange("list") 
+            }
         }
+
+    }
+
+    function onEdit(mail){
+        setComposedMail(mail)
+        setComposeChange(true)
+
+    }
+
+    async function onCompose(){
+        let email = emailService.newEmail()
+        email = await emailService.save(email)
+        setComposedMail(email)
+        setComposeChange(true)
+    }
+
+    async function onComposeChange(email){
+       setComposedMail(prev => ({...prev, ...email}))
+    }
+
+    async function onComposeFinish(com, email){
+        if(com === "edit") return setComposedMail(prev => ({...prev, ...email}))
+                    if(com === "close") {
+                        console.log("closed compose")
+                        await loadEmail()
+                        return setComposeChange(false)
+                    }
+            
+                    //on send
+                    await emailService.save(emailService.finalizeMail(email))
+                    setComposeChange(false)
+    }
+    async function saveDraft(){
+        await emailService.save(comMail)
     }
     
     if(!emails) return <div>Loading...</div>
@@ -94,10 +147,11 @@ export function EmailIndex() {
         <div className="logo">           
             <img src={imgUrl} alt="" />
             <h1>gMail</h1>
+            <button onClick={onCompose}>compose</button>
         </div>  
         <MailFilter filterBy={filterBy.filter} onFilterBy={onFilterBy}/>
         <Folders   filterBy={filterBy.folder} onFolderBy={onFolderBy}/>
-        <Test test={filterBy.filter} />
+        {compose && <Compose onFinish={onComposeFinish} onChange={onComposeChange} mail = {comMail}/>}
         <div className="mail-section">
             {mode === "list" && <MailList emails={emails} handleClick={onHandleClick} />}
             {mode === "deteils" && <MailDeteils handleClick={onHandleClick}/>}
@@ -106,7 +160,3 @@ export function EmailIndex() {
   
 }
     
-function Test({test}){
-    console.log("test",test)
-    return<div>test</div>
-}
